third_party/WebKit/LayoutTests/http/tests/serviceworker/serviceworkerglobalscope-scope.html

<!DOCTYPE html>
<title>ServiceWorkerGlobalScope: scope property</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/test-helpers.js"></script>
<script>

function scope_test(name, scope) {
    var t = async_test('Verify the scope property: ' + name);

    t.step(function() {
        var expectedScope, options;
        if (scope) {
            expectedScope = new URL(scope, document.location).toString();
        } else {
            expectedScope = new URL('/', document.location).toString();
        }

        service_worker_unregister_and_register(
            t, 'resources/serviceworkerglobalscope-scope-worker.js', scope
        ).then(
            t.step_func(function(worker) {
                var messageChannel = new MessageChannel();
                messageChannel.port1.onmessage = t.step_func(function(e) {
                    message = e.data;
                    assert_equals(message.initialScope, expectedScope, 'Worker should see the scope on eval.');
                    assert_equals(message.currentScope, expectedScope, 'Worker scope should not change.');
                    service_worker_unregister_and_done(t, scope);
                });
                worker.postMessage('', [messageChannel.port2]);
            })
        );
    });
}

scope_test('default');
scope_test('relative path', '/a/b/c/');
scope_test('absolute url', 'http://127.0.0.1:8000/');

</script>
